fixed routing and added Specials Component for testing
