+++ src/pages/Home.css
@@ -1,0 +1,17 @@
+.cardContainer {
+  display: flex;
+  margin: 5rem 3rem;
+  justify-content: space-between;
+}
+
+.wrapper {
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  gap: 1.5rem;
+}
+
+.home-container {
+  width: "100%";
+  justify-content: center;
+  align-items: center;
+}
