+++ src/pages/Specials.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+const Specials = () => {
+  return (
+    <div>
+      <h1>Specials</h1>
+    </div>
+  );
+};
+
+export default Specials;
