--- src/pages/index.jsx
@@ -1,18 +1,0 @@
-import React from "react";
-
-const Home = () => {
-  return (
-    <div
-      style={{
-        display: "flex",
-        justifyContent: "center",
-        alignItems: "center",
-        height: "90vh",
-      }}
-    >
-      <h1>Home</h1>
-    </div>
-  );
-};
-
-export default Home;
